test: added unit test
This test affirms that ashley does indeed work at npm.

As the original module only did a console log I opted to change the api to have index return true.

It would be possible to use spawn to test only the console log and not include the breaking api change.

It is worth noting that this is a semver major change